refactor(user): :recycle: update return types in domain services to UserAggregate and ListUserAggregate for improved consistency
