+++ src/modules/user/domain/domain-services/destroy-user.domain-service.ts
@@ -1,23 +1,23 @@
 import DomainException from '@common/domain/exceptions/domain.exception';
+import { UserAggregate } from '@user/domain/aggregates/user.aggregate';
 import { IUserRepositoryContract } from '@user/domain/contracts/user-repository.contract';
 import { ExceptionFactory } from '@user/domain/exceptions/exception.factory';
-import { UserNotFoundException } from '@user/domain/exceptions/user-not-found.exceptions';
-import { UserModel } from '@user/domain/models/user/user.model';
+import { UserNotFoundException } from '@user/domain/exceptions/user-not-found.exception';
 
 export class DestroyUserDomainService {
   constructor(private readonly userRepository: IUserRepositoryContract) {}
 
-  async go(uuid: string): Promise<UserModel> {
+  async go(uuid: string): Promise<UserAggregate> {
     try {
-      const userModel = await this.userRepository.getOneBy(uuid);
+      const userAggregate = await this.userRepository.getOneBy(uuid);
 
-      if (!userModel) {
+      if (!userAggregate) {
         throw new UserNotFoundException(`User with uuid ${uuid} not found`);
       }
 
-      userModel.destroy(userModel.uuid);
+      userAggregate.destroy();
 
-      return userModel;
+      return userAggregate;
     } catch (error) {
       if (error instanceof DomainException) {
         ExceptionFactory.createException(error.name, error.message);
