+++ src/modules/user/domain/domain-services/get-user.domain-service.ts
@@ -1,12 +1,13 @@
 import DomainException from '@common/domain/exceptions/domain.exception';
+import { UserAggregate } from '@user/domain/aggregates/user.aggregate';
 import { IUserRepositoryContract } from '@user/domain/contracts/user-repository.contract';
 import { ExceptionFactory } from '@user/domain/exceptions/exception.factory';
-import { UserNotFoundException } from '@user/domain/exceptions/user-not-found.exceptions';
+import { UserNotFoundException } from '@user/domain/exceptions/user-not-found.exception';
 
 export class GetUserDomainService {
   constructor(private readonly userRepository: IUserRepositoryContract) {}
 
-  async go(uuid: string) {
+  async go(uuid: string): Promise<UserAggregate> {
     try {
       const user = await this.userRepository.getOneBy(uuid);
