+++ src/modules/user/domain/domain-services/list-user.domain-service.ts
@@ -1,22 +1,22 @@
-import { ExceptionFactory } from '@user/domain/exceptions/exception.factory';
-import { UserNotFoundException } from '@user/domain/exceptions/user-not-found.exception';
-
 import { Criteria } from '@common/domain/criteria/criteria';
 import DomainException from '@common/domain/exceptions/domain.exception';
+import { ListUserAggregate } from '@user/domain/aggregates/list-user.aggregate';
 import { IUserRepositoryContract } from '@user/domain/contracts/user-repository.contract';
+import { ExceptionFactory } from '@user/domain/exceptions/exception.factory';
+import { UserNotFoundException } from '@user/domain/exceptions/user-not-found.exception';
 
 export class ListUserDomainService {
   constructor(private readonly userRepository: IUserRepositoryContract) {}
 
-  async go(criteria: Criteria): Promise<ListUserModel> {
+  async go(criteria: Criteria): Promise<ListUserAggregate> {
     try {
-      const users = await this.userRepository.matching(criteria);
+      const listUserAggregate = await this.userRepository.matching(criteria);
 
-      if (users.getTotal === 0) {
+      if (listUserAggregate.getTotal === 0) {
         throw new UserNotFoundException('No users found');
       }
 
-      return users;
+      return listUserAggregate;
     } catch (error) {
       if (error instanceof DomainException) {
         ExceptionFactory.createException(error.name, error.message);
